refactor: moved message logic into MessageUtil
