+++ Images-Core/src/main/java/com/andavin/images/Images.java
@@ -39,7 +39,6 @@ import org.apache.commons.lang.StringUtils;
 import org.bukkit.Bukkit;
 import org.bukkit.Location;
 import org.bukkit.configuration.file.FileConfiguration;
-import org.bukkit.configuration.file.YamlConfiguration;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.EventPriority;
@@ -90,7 +89,6 @@ public class Images extends JavaPlugin implements Listener {
     private static Images instance;
     private static File imagesDirectory;
     private static DataManager dataManager;
-    private static YamlConfiguration messages;
     private static final List<CustomImage> IMAGES = new ArrayList<>();
     private static final Map<UUID, Long> LAST_MOVE_TIMES = new HashMap<>();
     private static final PacketListener BRIDGE = Versioned.getInstance(PacketListener.class);
@@ -99,9 +97,9 @@ public class Images extends JavaPlugin implements Listener {
     @Override
     public void onLoad() {
         instance = this;
-        Images.loadMessages();
-        Logger.initialize(this.getLogger());
         imagesDirectory = this.getDataFolder();
+        MessageUtil.loadMessages(this, imagesDirectory);
+        Logger.initialize(this.getLogger());
         PacketListener.getImages = () -> IMAGES;
         setFieldValue(Scheduler.class, null, "instance", this);
         setFieldValue(TimeoutMetadata.class, null, "instance", this);
@@ -275,27 +273,6 @@ public class Images extends JavaPlugin implements Listener {
                 from.getBlockZ() >> 4 != to.getBlockZ() >> 4) {
             Scheduler.async(() -> this.refreshImages(player, to));
         }
-    }
-
-    /**
-     * Load messages from the messages.yml file.
-     */
-    private static void loadMessages() {
-        File messagesFile = new File(imagesDirectory, "messages.yml");
-        if (!messagesFile.exists()) {
-            instance.saveResource("messages.yml", false);
-        }
-        messages = YamlConfiguration.loadConfiguration(messagesFile);
-    }
-
-    /**
-     * Get a message from the messages.yml file.
-     *
-     * @param key The message key.
-     * @return The message string, or the key if not found.
-     */
-    public static String getMessage(String key) {
-        return messages.getString(key, key);
     }
 
     /**
